Fix main.cpp and test compilation errors
main.cpp fixes:
- Change CliParser::parse() to parseArguments() (free function, not class method)
- Change CliParser::getHelpMessage() to generateHelpMessage()
- Change CliParser::getVersionString() to generateVersionString()

CpuMonitorTest.cpp fixes:
- Cast uint64_t to double in EXPECT_NEAR to avoid C4244 warning
- Add explicit .0 to tolerance value for clarity

These changes fix compilation errors on Windows MSVC with /WX flag.

+++ tests/CpuMonitorTest.cpp
@@ -178,7 +178,7 @@ TEST_F(CpuMonitorTest, AverageFrequencyCalculatedCorrectly) {
     uint64_t expectedAvg = sum / stats.cores.size();
     
     // Should be within 10 MHz of calculated average (allow for rounding)
-    EXPECT_NEAR(stats.averageFrequencyMhz, expectedAvg, 10);
+    EXPECT_NEAR(static_cast<double>(stats.averageFrequencyMhz), static_cast<double>(expectedAvg), 10.0);
 }
 
 // Test 11: Usage under load increases (stress test)
